Move math functions out of awful.util into new gears.math
