+++ lib/gears/init.lua
@@ -18,6 +18,7 @@ return
     shape = require("gears.shape");
     protected_call = require("gears.protected_call");
     geometry = require("gears.geometry");
+    math = require("gears.math");
 }
 
 -- vim: filetype=lua:expandtab:shiftwidth=4:tabstop=8:softtabstop=4:textwidth=80
